Task1-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.8",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
